+++ src/fonduer/meta.py
@@ -2,7 +2,6 @@ import logging
 from builtins import object
 from urllib.parse import urlparse
 
-import psycopg2
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
@@ -16,25 +15,27 @@ def new_sessionmaker():
     # http://oddbird.net/2014/06/14/sqlalchemy-postgres-autocommit/
     # Otherwise any e.g. query starts a transaction, locking tables... very
     # bad for e.g. multiple notebooks open, multiple processes, etc.
-    if Meta.postgres and Meta.ready:
+    try:
         engine = create_engine(
             Meta.conn_string,
             client_encoding="utf8",
             use_batch_mode=True,
             isolation_level="AUTOCOMMIT",
         )
-    else:
+        engine.connect()
+    except Exception:
         raise ValueError(
             "Meta variables have not been initialized with "
             "a valid postgres connection string."
+            "Use the form: "
+            "postgresql://<user>:<pw>@<host>:<port>/<database_name>"
         )
     # New sessionmaker
     return sessionmaker(bind=engine)
 
 
-def _validate_conn_string(conn_string):
-    """Check that the PostgreSQL connection string is valid."""
-    logger.debug("Validating {} as a connection string...".format(conn_string))
+def _update_meta(conn_string):
+    """Update Meta class."""
     url = urlparse(conn_string)
     Meta.conn_string = conn_string
     Meta.DBNAME = url.path[1:]
@@ -42,23 +43,7 @@ def _validate_conn_string(conn_string):
     Meta.DBPWD = url.password
     Meta.DBHOST = url.hostname
     Meta.DBPORT = url.port
-    Meta.postgres = url.scheme.startswith("postgres")
-    # Actually try to connect to see if the connection string is valid
-    try:
-        con = psycopg2.connect(
-            database=Meta.DBNAME,
-            user=Meta.DBUSER,
-            password=Meta.DBPWD,
-            host=Meta.DBHOST,
-        )
-        con.close()
-    except psycopg2.OperationalError as e:
-        raise ValueError(
-            "{} is an invalid connection string. Use the form {}".format(
-                conn_string, "postgres://<user>:<pw>@<host>:<port>/<database_name>"
-            )
-        )
-    return True
+    Meta.postgres = url.scheme.startswith("postgresql")
 
 
 class Meta(object):
@@ -79,13 +64,12 @@ class Meta(object):
     engine = None
     Base = declarative_base(name="Base", cls=object)
     postgres = False
-    ready = False
 
     @classmethod
     def init(cls, conn_string=None):
         """Return the unique Meta class."""
         if conn_string:
-            Meta.ready = _validate_conn_string(conn_string)
+            _update_meta(conn_string)
             # We initialize the engine within the models module because models'
             # schema can depend on which data types are supported by the engine
             Meta.Session = new_sessionmaker()